requirements.txt
boto==2.42.0
Django==1.10.2
django-storages==1.5.1
psycopg2==2.6.2
wheel==0.24.0
